Add ElasticSearch resource check to Canary
Summary:
prerequisite for T4115

Adds an ElasticSearch check to Canary. At first, I wrote it with urllib2 to use the standard library, but then I saw that `requests` was in `install_requires` so I converted to that. And then discovered that mocking requests meant you couldn't test more than one mock at a time. So I rewrote the tests to use dcramer's `responses`. It's a little awkward because I shoehorned it into the existing tests. I wrote more unit-test-y tests for the elasticsearch stuff in `test_search_endpoint.py` that the other endpoints don't have.

TODO before landing:
* bump minor version

Possible things to do in the future:
* add equivalent functionality to rCANSCALA
* add ability to get info from Haystack + ElasticSearch

Test Plan:
* `pip install -r requirements.txt`
* `make test`
* `tox` to get coverage over multiple versions of django
* I also hooked it up to rYCAKE in a local branch to accomplish T4115

Reviewers: scott.sullivan, noah

Reviewed By: noah

Subscribers: noah, raul, niran, jeff.israel, philip.jameson, dennishdoire, whitney.reeds, #canary

Differential Revision: https://phab.tabbedout.com/D3078